Long Money update + replaced PGMoney with Long in Java, money with int in SQL + added lots of tests P.S.: I think 90% of work is done... maybe...

+++ src/main/resources/sql/film_create.sql
@@ -7,6 +7,6 @@ CREATE TABLE film (
     company text NOT NULL,
     director text NOT NULL,
     release_year int NOT NULL,
-    purchase_price money,
-    rent_price money
+    purchase_price int,
+    rent_price int
 );
